requirements.txt
fastapi>=0.109,<1
uvicorn[standard]>=0.27
gunicorn>=22.0
jinja2>=3.1
starlette>=0.36
pydantic>=2.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
